\name{internal}
\alias{internal}


\alias{linestMatrix}
\alias{linestMatrix.default}

\alias{print.linestMatrix}
\alias{print.linest}

\alias{print.LSmeans}

\alias{pdiff}


\title{Internal functions for the 'linest' package}
\description{Internal functions for the 'doBy' package}
\keyword{internal}% at least one, from doc/KEYWORDS
